Made a functional of create list of the word and fixed user auth
Improved the authentication process for the user by moving its processing to the UserAuthHandler. Implemented the functionality for creating a new list of words in the WordListHandler. Created a new enum for the characteristics of the list of words, and a new method in the KeyboardBuilder for creating a keyboard of the elements of this enum. Added new properties to the application.properties file to resolve issues with lazy initialization. These changes should enhance the overall functionality of the code.