Added functionality for showing all word lists, added new test type, and improved Test model.
Improved functionality by adding the ability to show all word lists and their corresponding information. Also added a new condition to the test conditions and included new fields in the Test model: startTime, finishTime, and spendTime. Created a utility class for formatting time into a more user-friendly format. Furthermore, a new button 📊 All tests was added to the functional keyboard. Additionally, a new type of test was created to check the correctness of word writing.