Refactor processing of the added a new lists.
Moved a functional of the suggested the lists of the words. Also added an `@EqualsAndHashCode` to the Test and Word. Changed a method of the get suggested lists of the words, dded a check of the access of the level and id of the user. Also fixed the added a methods for users.