Added Notification model, simplified keyboards and refactored code
Created a Notification model, along with related enums for the model, repository, service, and empty handler. Simplified the source keyboard and added a new keyboard that appears when the user selects a point in the main keyboard. The new keyboard is generated by methods that create keyboards from lists of strings. Added lists of button titles as private fields to KeyboardBuilder for each keyboard, to simplify code and improve readability.